+++ src/hipay_enterprise/controllers/front/redirect.php
@@ -131,9 +131,9 @@ class Hipay_enterpriseRedirectModuleFrontController extends ModuleFrontControlle
                     );
 
                     $path = 'payment/ps16/paymentFormApi-16.tpl';
-                    $this->addJS(array(_MODULE_DIR_ . 'hipay_enterprise/views/js/strings.js'));
-                    $this->addJS(array(_MODULE_DIR_ . 'hipay_enterprise/views/js/card-js.min.js'));
                 }
+                $this->addJS(array(_MODULE_DIR_ . 'hipay_enterprise/views/js/strings.js'));
+                $this->addJS(array(_MODULE_DIR_ . 'hipay_enterprise/views/js/card-js.min.js'));
                 break;
             default:
                 break;