+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+# Version 2.1.4
+
+- Fix: credit card block are displaying even if no credit card payment are activated
+
+# Version 2.1.3
+
+- Fix local payment with hosted order
+
+# Version 2.1.2
+
+- Fix translations
+
+# Version 2.1.1
+
+- Fix error 500 on refused notification
+
+# Version 2.1.0
+
+- Fix link in PrestaShop BO (Module configuration page)
+- Add upgrade script
+- Fix Oneclick payment bug on 1.6 and 1.7
+
+# Version 2.0.5
+
+- Add payment method **Bnp personal Finance**
+- Add log for concurrent transactions
+- Override "OrderExists" method for concurrent notifications (Bug with prestashop cache)  
+- Add translation support for payment method name in front-office. 
+- Add FAQ translation
+- Change form handling for passphrase
+
 # Version 2.0.4
 
 - Fix Product key for Prestashop addons